refactor(subsequences): move survey helpers out of the utils package init DEV-2291
